requirements.txt
mkl
numpy
six
psutil
future
pygtail
pyusb
requests
pexpect
seabreeze
pyudev
pyserial
wasatch
pywin32
crcmod
bleak
matplotlib
scikit-learn
pyftdi